# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/shifts": {
      "post": {
        "operationId": "ShiftsController_create",
        "parameters": [
          {
            "name": "X-Organisation-Slug",
            "in": "header",
            "description": "Slug of the organisation the request acts on. The caller must be a member of it. Required — the request is rejected without it.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-logistics"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShiftDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Empty planned shift created. This consumes one shift from the organisation allowance.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing `X-Organisation-Slug` header, or the request body failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or expired bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "402": {
            "description": "The organisation is over its free shift allowance and has no payment method on file.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Unknown organisation, the caller is not a member of it, or the required permission is missing.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "That driver or vehicle already has an open shift on this date.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "shifts"
        ]
      }
    },
...
    "/api/v1/shifts/{id}/dispatch": {
      "post": {
        "operationId": "ShiftsController_dispatch",
        "parameters": [
          {
            "name": "X-Organisation-Slug",
            "in": "header",
            "description": "Slug of the organisation the request acts on. The caller must be a member of it. Required — the request is rejected without it.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-logistics"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Moves the shift planned to dispatched, closing it to automatic assignment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing `X-Organisation-Slug` header, or the request body failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or expired bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Unknown organisation, the caller is not a member of it, or the required permission is missing.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "The shift is not in the planned state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "shifts"
        ]
      }
    },
    "/api/v1/shifts/{id}/packages": {
      "post": {
        "operationId": "ShiftsController_addPackages",
        "parameters": [
          {
            "name": "X-Organisation-Slug",
            "in": "header",
            "description": "Slug of the organisation the request acts on. The caller must be a member of it. Required — the request is rejected without it.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-logistics"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPackagesToShiftDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dispatcher override: pins the packages to this shift and returns the rewritten plan. Feasibility still runs, but a breach is reported as a warning rather than refused.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftPlanDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing `X-Organisation-Slug` header, or the request body failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or expired bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Unknown organisation, the caller is not a member of it, or the required permission is missing.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "The shift is already dispatched.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "shifts"
        ]
      }
    },
    "/api/v1/shifts/{id}/packages/{packageId}": {
      "delete": {
        "operationId": "ShiftsController_removePackage",
        "parameters": [
          {
            "name": "X-Organisation-Slug",
            "in": "header",
            "description": "Slug of the organisation the request acts on. The caller must be a member of it. Required — the request is rejected without it.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-logistics"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "packageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Removes the package and returns the rewritten plan. The route steps are deleted and re-inserted in order rather than renumbered.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftPlanDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing `X-Organisation-Slug` header, or the request body failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or expired bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Unknown organisation, the caller is not a member of it, or the required permission is missing.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "The package is loaded, in transit or already delivered, so it cannot be removed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "shifts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hikyaku Logistics API",
    "description": "Backend for the Hikyaku dashboard and the public booking site.\n\n**Authentication.** Every route except the public booking and routing endpoints requires a Supabase access token as `Authorization: Bearer <jwt>`.\n\n**Tenancy.** Authenticated routes resolve the active organisation from the `X-Organisation-Slug` header. The unauthenticated booking and routing routes read `x-org-slug` instead — a separate header, not a casing variant. Each operation declares the one it expects.\n\n**Errors.** Failures use Nest’s default body (`{ statusCode, message, error }`); `message` is an array of strings when request validation is what failed. See `ApiErrorDto`.",
    "version": "1.0",
    "contact": {
      "name": "Hikyaku Engineering",
      "url": "https://hikyaku.org",
      "email": "engineering@hikyaku.org"
    },
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    }
  },
...
  "servers": [
    {
      "url": "https://api.hikyaku.org",
      "description": "Production"
    },
    {
      "url": "http://localhost:3002",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "Mirrors the HTTP status code of the response.",
            "example": 400
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "description": "A single message for a hand-thrown exception, or one entry per failed constraint when request validation is what rejected the call.",
            "example": "Missing X-Organisation-Slug header"
          },
          "error": {
            "type": "string",
            "description": "HTTP reason phrase for the status code.",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "CreateShiftDto": {
        "type": "object",
        "properties": {
          "warehouseId": {
            "type": "string",
            "format": "uuid",
            "description": "warehouse.id the shift starts and ends at."
          },
          "driverId": {
            "type": "string",
            "format": "uuid",
            "description": "drivers.id. Must share a warehouse with vehicleId."
          },
          "vehicleId": {
            "type": "string",
            "format": "uuid",
            "description": "vehicles.id. Also resolves the routing profile via vehicles.vehicle_type."
          },
          "shiftDate": {
            "type": "string",
            "format": "date",
            "example": "2026-09-01",
            "description": "Warehouse-local service day (YYYY-MM-DD). A driver or vehicle can hold at most one open shift per day."
          },
          "scheduledStart": {
            "type": "string",
            "format": "date-time",
            "description": "When the vehicle sets off. Omitted, the shift stays open to automatic assignment indefinitely; set, it closes 15 minutes before this time so nothing is added to a van about to roll."
          }
        },
        "required": [
          "warehouseId",
          "driverId",
          "vehicleId",
          "shiftDate"
        ]
      },
      "ShiftDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "vrp_optimization.id."
          },
          "status": {
            "type": "string",
            "enum": [
              "planned",
              "dispatched",
              "completed",
              "cancelled"
            ],
            "description": "`planned` is the only state open to automatic assignment, and only until 15 minutes before scheduledStart."
          },
          "organisationId": {
            "type": "string",
            "format": "uuid"
          },
          "warehouseId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "driverId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "vehicleId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "shiftDate": {
            "type": "string",
            "format": "date"
          },
          "scheduledStart": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "routeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "vrp_route.id of the single route, or null while the shift is empty."
          },
          "stopCount": {
            "type": "number",
            "description": "Job steps on the route, excluding depot start/end."
          },
          "revision": {
            "type": "number",
            "description": "Bumped on every plan rewrite."
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "status",
          "organisationId",
          "warehouseId",
          "driverId",
          "vehicleId",
          "shiftDate",
          "scheduledStart",
          "routeId",
          "stopCount",
          "revision",
          "updatedAt"
        ]
      },
...
      "AddPackagesToShiftDto": {
        "type": "object",
        "properties": {
          "packageIds": {
            "description": "Existing packages.id values. Candidate selection is bypassed, but feasibility still runs: a package that breaks a deadline is reported as a warning rather than refused.",
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "packageIds"
        ]
      },
      "ShiftPackageOutcomeDto": {
        "type": "object",
        "properties": {
          "packageId": {
            "type": "string",
            "format": "uuid"
          },
          "added": {
            "type": "boolean",
            "description": "False when the package was already claimed elsewhere."
          },
          "warning": {
            "type": "string",
            "nullable": true,
            "description": "Set when the package was added but breaks something — e.g. its own deadline, or another stop’s. The dispatcher decided; we record it."
          }
        },
        "required": [
          "packageId",
          "added"
        ]
      },
      "ShiftPlanDto": {
        "type": "object",
        "properties": {
          "shift": {
            "$ref": "#/components/schemas/ShiftDto"
          },
          "packages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShiftPackageOutcomeDto"
            }
          }
        },
        "required": [
          "shift",
          "packages"
        ]
      },
...
    }
  }
}